Simplify handlers module organization
