+++ aries_vcx/src/handlers/mod.rs
@@ -2,10 +2,10 @@ use futures::future::BoxFuture;
 
 use crate::error::prelude::*;
 use crate::handlers::connection::connection::ConnectionState;
-use crate::handlers::issuance::holder::holder::HolderState;
-use crate::handlers::issuance::issuer::issuer::IssuerState;
-use crate::handlers::proof_presentation::prover::prover::ProverState;
-use crate::handlers::proof_presentation::verifier::verifier::VerifierState;
+use crate::handlers::issuance::holder::HolderState;
+use crate::handlers::issuance::issuer::IssuerState;
+use crate::handlers::proof_presentation::prover::ProverState;
+use crate::handlers::proof_presentation::verifier::VerifierState;
 use crate::messages::a2a::A2AMessage;
 use crate::protocols::connection::invitee::state_machine::InviteeState;
 use crate::protocols::connection::inviter::state_machine::InviterState;
