--- aries_vcx/src/handlers/out_of_band/receiver/mod.rs
@@ -1,1 +1,0 @@
-pub mod receiver;
