--- aries_vcx/src/handlers/out_of_band/sender/mod.rs
@@ -1,1 +1,0 @@
-pub mod sender;
