--- aries_vcx/src/handlers/proof_presentation/verifier/mod.rs
@@ -1,1 +1,0 @@
-pub mod verifier;
