+++ aries_vcx/src/protocols/issuance/issuer/state_machine.rs
@@ -2,7 +2,7 @@ use std::collections::HashMap;
 use std::fmt::Display;
 
 use crate::error::{VcxError, VcxErrorKind, VcxResult};
-use crate::handlers::issuance::issuer::issuer::IssuerState;
+use crate::handlers::issuance::issuer::IssuerState;
 use crate::libindy::credentials::encode_attributes;
 use crate::handlers::SendClosure;
 use crate::libindy::utils::anoncreds;
