+++ aries_vcx/tests/test_integration.rs
@@ -87,15 +87,15 @@ mod tests {
     use aries_vcx::error::VcxResult;
     use aries_vcx::handlers::connection::connection::{Connection, ConnectionState};
     use aries_vcx::handlers::issuance::holder::test_utils::get_credential_offer_messages;
-    use aries_vcx::handlers::issuance::holder::holder::{Holder, HolderState};
+    use aries_vcx::handlers::issuance::holder::{Holder, HolderState};
     use aries_vcx::handlers::issuance::issuer::test_utils::get_credential_proposal_messages;
-    use aries_vcx::handlers::issuance::issuer::issuer::{Issuer, IssuerConfig, IssuerState};
+    use aries_vcx::handlers::issuance::issuer::{Issuer, IssuerConfig, IssuerState};
     use aries_vcx::handlers::out_of_band::{GoalCode, HandshakeProtocol, OutOfBand};
-    use aries_vcx::handlers::out_of_band::receiver::receiver::OutOfBandReceiver;
-    use aries_vcx::handlers::out_of_band::sender::sender::OutOfBandSender;
+    use aries_vcx::handlers::out_of_band::receiver::OutOfBandReceiver;
+    use aries_vcx::handlers::out_of_band::sender::OutOfBandSender;
     use aries_vcx::handlers::proof_presentation::prover::test_utils::get_proof_request_messages;
-    use aries_vcx::handlers::proof_presentation::prover::prover::{Prover, ProverState};
-    use aries_vcx::handlers::proof_presentation::verifier::verifier::{Verifier, VerifierState};
+    use aries_vcx::handlers::proof_presentation::prover::{Prover, ProverState};
+    use aries_vcx::handlers::proof_presentation::verifier::{Verifier, VerifierState};
     use aries_vcx::libindy::credential_def::{CredentialDef, CredentialDefConfigBuilder, RevocationDetailsBuilder};
     use aries_vcx::libindy::proofs::proof_request_internal::{AttrInfo, NonRevokedInterval, PredicateInfo};
     use aries_vcx::libindy::utils::anoncreds::test_utils::create_and_write_test_schema;
