+++ aries_vcx/tests/utils/devsetup_agent.rs
@@ -5,11 +5,11 @@ pub mod test {
     use aries_vcx::handlers::connection::connection::{Connection, ConnectionState};
     use aries_vcx::handlers::connection::public_agent::PublicAgent;
     use aries_vcx::handlers::issuance::holder::test_utils::get_credential_offer_messages;
-    use aries_vcx::handlers::issuance::holder::holder::{Holder, HolderState};
-    use aries_vcx::handlers::issuance::issuer::issuer::{Issuer, IssuerState};
+    use aries_vcx::handlers::issuance::holder::{Holder, HolderState};
+    use aries_vcx::handlers::issuance::issuer::{Issuer, IssuerState};
     use aries_vcx::handlers::proof_presentation::prover::test_utils::get_proof_request_messages;
-    use aries_vcx::handlers::proof_presentation::prover::prover::{Prover, ProverState};
-    use aries_vcx::handlers::proof_presentation::verifier::verifier::{Verifier, VerifierState};
+    use aries_vcx::handlers::proof_presentation::prover::{Prover, ProverState};
+    use aries_vcx::handlers::proof_presentation::verifier::{Verifier, VerifierState};
     use aries_vcx::init::{create_agency_client_for_main_wallet, init_issuer_config, open_as_main_wallet};
     use aries_vcx::libindy::credential_def::{CredentialDef, CredentialDefConfigBuilder, RevocationDetails};
     use aries_vcx::libindy::credential_def::PublicEntityStateType;
